Prebaceno sa h2 na postgresql Prepravljeno za redis Dodat docker compose
