+++ banka-service/src/test/java/rs/edu/raf/realizacijaTransakcija/TransakcijaServisImplTest.java
@@ -11,7 +11,7 @@ import rs.edu.raf.model.entities.racun.DevizniRacun;
 import rs.edu.raf.model.entities.racun.PravniRacun;
 import rs.edu.raf.model.entities.racun.TekuciRacun;
 import rs.edu.raf.repository.transaction.*;
-import rs.edu.raf.service.ExchangeRateService;
+import rs.edu.raf.service.ExchangeRateServiceImpl;
 import rs.edu.raf.service.racun.RacunServis;
 import rs.edu.raf.model.entities.transaction.PrenosSredstava;
 import rs.edu.raf.model.entities.transaction.Status;
@@ -58,7 +58,7 @@ public class TransakcijaServisImplTest {
     private DevizniRacun devizniRacun;
 
     private PrenosSredstava prenosSredstava;
-    private ExchangeRateService exchangeRateService;
+    private ExchangeRateServiceImpl exchangeRateServiceImpl;
 
     private Uplata uplata;
 
@@ -73,7 +73,7 @@ public class TransakcijaServisImplTest {
                 pravniRacunRepository,
                 tekuciRacunRepository,
                 devizniRacunRepository,
-                exchangeRateService,
+                exchangeRateServiceImpl,
                 racunServis
         ));
     }
